Add support routines for async basic example
